+++ point.h
@@ -8,9 +8,9 @@ typedef struct point {
 	uint16_t col;
 } point;
 
-void transX(struct point *p, int16_t dist);
-void transY(struct point *p, int16_t dist);
-void transZ(struct point *p, int16_t dist);
+void transX(struct point *p, float dist);
+void transY(struct point *p, float dist);
+void transZ(struct point *p, float dist);
 void rotX(struct point *p, int16_t angle);
 void rotY(struct point *p, int16_t angle);
 void rotZ(struct point *p, int16_t angle);
